rename container components
