+++ src/app/components/Brand/index.stories.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
-import BrandContainer from '../../containers/BrandContainer/index';
+import BrandContainer from '../../containers/Brand';
 
 storiesOf('Brand', module).add('default', () => <BrandContainer />);
