--- src/app/containers/HeaderContainer/index.stories.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
-import HeaderContainer from './index';
-
-storiesOf('HeaderContainer', module).add('default', () => <HeaderContainer />);
